+++ src/builtin.h
@@ -2570,6 +2570,35 @@ ARSH_METHOD edit_configs(RuntimeContext &ctx) {
   RET(editor.getConfigs(ctx));
 }
 
+//!bind: function abbr($this: LineEditor, $pattern: String, $expansion: Option<String>): Void
+ARSH_METHOD edit_abbr(RuntimeContext &ctx) {
+  SUPPRESS_WARNING(edit_abbr);
+  auto &editor = typeAs<LineEditorObject>(LOCAL(0));
+  CHECK_EDITOR_LOCK(editor);
+  auto pattern = LOCAL(1).asStrRef();
+  StringRef expand;
+  if (!LOCAL(2).isInvalid()) {
+    expand = LOCAL(2).asStrRef();
+  }
+  editor.defineAbbr(ctx, pattern, expand);
+  RET_VOID;
+}
+
+//!bind: function abbrs($this: LineEditor): Map<String, String>
+ARSH_METHOD edit_abbrs(RuntimeContext &ctx) {
+  SUPPRESS_WARNING(edit_abbrs);
+  auto &editor = typeAs<LineEditorObject>(LOCAL(0));
+  auto &stringType = ctx.typePool.get(TYPE::String);
+  auto ret = ctx.typePool.createMapType(stringType, stringType);
+  assert(ret);
+  auto &mapType = cast<MapType>(*ret.asOk());
+  auto value = Value::create<OrderedMapObject>(mapType, ctx.getRng().next());
+  for (auto &[p, e] : editor.getAbbrMap()) {
+    typeAs<OrderedMapObject>(value).insert(Value::createStr(p), Value::createStr(e));
+  }
+  RET(value);
+}
+
 //!bind: function name($this : CLI) : String
 ARSH_METHOD cli_get(RuntimeContext &ctx) {
   SUPPRESS_WARNING(cli_get);